+++ test/TestInstructions.cpp
@@ -49,19 +49,19 @@ void TestInstructions::testConditionCodes()
 
 void TestInstructions::testConstantSaturations()
 {
-	TEST_ASSERT_EQUALS(127l, saturate<char>(1024));
-	TEST_ASSERT_EQUALS(255l, saturate<unsigned char>(1024));
-	TEST_ASSERT_EQUALS(32767l, saturate<short>(100000));
-	TEST_ASSERT_EQUALS(65535l, saturate<unsigned short>(100000));
-	TEST_ASSERT_EQUALS(2147483647l, saturate<int>(static_cast<int64_t>(1) << 40));
-	TEST_ASSERT_EQUALS(4294967295l, saturate<unsigned int>(static_cast<int64_t>(1) << 40));
+	TEST_ASSERT_EQUALS(static_cast<int64_t>(127), saturate<signed char>(1024));
+	TEST_ASSERT_EQUALS(static_cast<int64_t>(255), saturate<unsigned char>(1024));
+	TEST_ASSERT_EQUALS(static_cast<int64_t>(32767), saturate<short>(100000));
+	TEST_ASSERT_EQUALS(static_cast<int64_t>(65535), saturate<unsigned short>(100000));
+	TEST_ASSERT_EQUALS(static_cast<int64_t>(2147483647), saturate<int>(static_cast<int64_t>(1) << 40));
+	TEST_ASSERT_EQUALS(static_cast<int64_t>(4294967295), saturate<unsigned int>(static_cast<int64_t>(1) << 40));
 
-	TEST_ASSERT_EQUALS(-128l, saturate<char>(-1024));
-	TEST_ASSERT_EQUALS(0l, saturate<unsigned char>(-1024));
-	TEST_ASSERT_EQUALS(-32768l, saturate<short>(-100000));
-	TEST_ASSERT_EQUALS(0l, saturate<unsigned short>(-100000));
-	TEST_ASSERT_EQUALS(-2147483648l, saturate<int>(-(static_cast<int64_t>(1) << 40)));
-	TEST_ASSERT_EQUALS(0l, saturate<unsigned int>(-(static_cast<int64_t>(1) << 40)));
+	TEST_ASSERT_EQUALS(static_cast<int64_t>(-128), saturate<signed char>(-1024));
+	TEST_ASSERT_EQUALS(static_cast<int64_t>(0), saturate<unsigned char>(-1024));
+	TEST_ASSERT_EQUALS(static_cast<int64_t>(-32768), saturate<short>(-100000));
+	TEST_ASSERT_EQUALS(static_cast<int64_t>(0), saturate<unsigned short>(-100000));
+	TEST_ASSERT_EQUALS(static_cast<int64_t>(-2147483648), saturate<int>(-(static_cast<int64_t>(1) << 40)));
+	TEST_ASSERT_EQUALS(static_cast<int64_t>(0), saturate<unsigned int>(-(static_cast<int64_t>(1) << 40)));
 }
 
 struct TestBitfield : public Bitfield<int32_t>
